backsrc/requirements.txt
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
boto3==1.9.158
botocore==1.12.158
certifi==2019.6.16
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
dj-database-url==0.5.0
django==2.2
django-cors-headers==3.4.0
django-filter==2.0.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-storages==1.7.1
django-tinymce==2.8.0
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
docutils==0.14
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
Pillow==6.1.0
psycopg2==2.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.22.0
s3transfer==0.2.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.3
whitenoise==4.1.2
zope.interface==5.1.0
